requirements.txt
flask
gunicorn
flask_sqlalchemy
opencv-contrib-python
pytesseract
numpy
imutils
Pillow
matplotlib
know_your_plates
tesseract
